+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -14,6 +14,8 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import roomescape.member.domain.Member;
+import roomescape.theme.domain.Theme;
+import roomescape.reservation.time.domain.ReservationTime;
 
 @Entity
 public class Reservation {
