+++ src/main/java/roomescape/reservation/dto/response/ReservationTimeResponse.java
@@ -4,7 +4,7 @@ import java.time.LocalTime;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
-import roomescape.reservation.domain.ReservationTime;
+import roomescape.reservation.time.domain.ReservationTime;
 
 public record ReservationTimeResponse(
         Long id,
