+++ src/main/java/roomescape/reservation/dto/response/WaitingResponse.java
@@ -3,7 +3,8 @@ package roomescape.reservation.dto.response;
 import java.time.LocalDate;
 
 import roomescape.member.dto.MemberResponse;
-import roomescape.reservation.domain.Waiting;
+import roomescape.reservation.waiting.domain.Waiting;
+import roomescape.theme.dto.response.ThemeResponse;
 
 public record WaitingResponse(
         Long id,
