+++ src/main/java/roomescape/reservation/repository/ReservationRepository.java
@@ -9,8 +9,8 @@ import org.springframework.data.jpa.repository.Query;
 import roomescape.member.domain.MemberId;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationId;
-import roomescape.reservation.domain.ReservationTimeId;
-import roomescape.reservation.domain.ThemeId;
+import roomescape.reservation.time.domain.ReservationTimeId;
+import roomescape.theme.domain.ThemeId;
 
 public interface ReservationRepository extends JpaRepository<Reservation, ReservationId> {
