+++ src/main/java/roomescape/reservation/waiting/repository/WaitingRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.repository;
+package roomescape.reservation.waiting.repository;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -8,13 +8,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 import roomescape.member.domain.MemberId;
-import roomescape.reservation.domain.ReservationTime;
-import roomescape.reservation.domain.ReservationTimeId;
-import roomescape.reservation.domain.Theme;
-import roomescape.reservation.domain.ThemeId;
-import roomescape.reservation.domain.Waiting;
-import roomescape.reservation.domain.WaitingId;
-import roomescape.reservation.domain.WaitingWithRank;
+import roomescape.reservation.time.domain.ReservationTime;
+import roomescape.reservation.time.domain.ReservationTimeId;
+import roomescape.theme.domain.Theme;
+import roomescape.theme.domain.ThemeId;
+import roomescape.reservation.waiting.domain.Waiting;
+import roomescape.reservation.waiting.domain.WaitingId;
+import roomescape.reservation.waiting.domain.WaitingWithRank;
 
 public interface WaitingRepository extends JpaRepository<Waiting, WaitingId> {
 
@@ -41,7 +41,7 @@ public interface WaitingRepository extends JpaRepository<Waiting, WaitingId> {
     );
 
     @Query("""
-            SELECT new roomescape.reservation.domain.WaitingWithRank(
+            SELECT new roomescape.reservation.waiting.domain.WaitingWithRank(
                 w,
                 (SELECT COUNT(w2)
                  FROM Waiting w2
