+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -29,21 +29,21 @@ import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
 import roomescape.member.repository.MemberRepository;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.domain.ReservationTime;
-import roomescape.reservation.domain.Theme;
-import roomescape.reservation.domain.Waiting;
+import roomescape.reservation.time.domain.ReservationTime;
+import roomescape.theme.domain.Theme;
+import roomescape.reservation.waiting.domain.Waiting;
 import roomescape.reservation.dto.request.ReservationCreateRequest;
 import roomescape.reservation.dto.response.BookedReservationTimeResponse;
 import roomescape.reservation.dto.response.ReservationResponse;
 import roomescape.reservation.dto.response.ReservationTimeResponse;
-import roomescape.reservation.dto.response.ThemeResponse;
+import roomescape.theme.dto.response.ThemeResponse;
 import roomescape.reservation.payment.dto.request.PaymentRequest;
 import roomescape.reservation.payment.repository.PaymentRepository;
 import roomescape.reservation.payment.service.PaymentService;
 import roomescape.reservation.repository.ReservationRepository;
-import roomescape.reservation.repository.ReservationTimeRepository;
-import roomescape.reservation.repository.ThemeRepository;
-import roomescape.reservation.repository.WaitingRepository;
+import roomescape.reservation.time.repository.ReservationTimeRepository;
+import roomescape.theme.repository.ThemeRepository;
+import roomescape.reservation.waiting.repository.WaitingRepository;
 
 @ActiveProfiles("test")
 @DataJpaTest
